aws-cost-optimizer/requirements.txt
# AWS
boto3>=1.28.0
botocore>=1.31.0

# Dynatrace
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Excel Generation
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
